+++ src/CinemAPI.Models/Ticket.cs
@@ -39,7 +39,7 @@ namespace CinemAPI.Models
         public int RoomNumber { get; set; }
         public int Row { get; set; }
         public int Col { get; set; }
-        public int ProjectionId { get; set; }
+        public long ProjectionId { get; set; }
         public int MovieId { get; set; }
     }
 }
